+++ src/actions/public/call.ts
@@ -9,10 +9,6 @@ import type { Client } from '../../clients/createClient.js'
 import type { Transport } from '../../clients/transports/createTransport.js'
 import { multicall3Abi } from '../../constants/abis.js'
 import { aggregate3Signature } from '../../constants/contract.js'
-import {
-  InvalidAddressError,
-  type InvalidAddressErrorType,
-} from '../../errors/address.js'
 import { BaseError } from '../../errors/base.js'
 import {
   ChainDoesNotSupportContract,
@@ -22,27 +18,12 @@ import {
   RawContractError,
   type RawContractErrorType,
 } from '../../errors/contract.js'
-import {
-  InvalidBytesLengthError,
-  type InvalidBytesLengthErrorType,
-} from '../../errors/data.js'
-import {
-  AccountStateConflictError,
-  type AccountStateConflictErrorType,
-  StateAssignmentConflictError,
-  type StateAssignmentConflictErrorType,
-} from '../../errors/stateOverride.js'
 import type { ErrorType } from '../../errors/utils.js'
 import type { BlockTag } from '../../types/block.js'
 import type { Chain } from '../../types/chain.js'
 import type { Hex } from '../../types/misc.js'
-import type {
-  RpcAccountStateOverride,
-  RpcStateMapping,
-  RpcStateOverride,
-  RpcTransactionRequest,
-} from '../../types/rpc.js'
-import type { StateMapping, StateOverride } from '../../types/stateOverride.js'
+import type { RpcTransactionRequest } from '../../types/rpc.js'
+import type { StateOverride } from '../../types/stateOverride.js'
 import type { TransactionRequest } from '../../types/transaction.js'
 import type { ExactPartial, UnionOmit } from '../../types/utils.js'
 import {
@@ -53,7 +34,6 @@ import {
   type EncodeFunctionDataErrorType,
   encodeFunctionData,
 } from '../../utils/abi/encodeFunctionData.js'
-import { isAddress } from '../../utils/address/isAddress.js'
 import type { RequestErrorType } from '../../utils/buildRequest.js'
 import {
   type GetChainContractAddressErrorType,
@@ -77,6 +57,10 @@ import {
   type CreateBatchSchedulerErrorType,
   createBatchScheduler,
 } from '../../utils/promise/createBatchScheduler.js'
+import {
+  type SerializeStateOverrideErrorType,
+  serializeStateOverride,
+} from '../../utils/stateOverride.js'
 import { assertRequest } from '../../utils/transaction/assertRequest.js'
 import type {
   AssertRequestErrorType,
@@ -113,7 +97,7 @@ export type CallReturnType = { data: Hex | undefined }
 
 export type CallErrorType = GetCallErrorReturnType<
   | ParseAccountErrorType
-  | ParseStateOverrideErrorType
+  | SerializeStateOverrideErrorType
   | AssertRequestErrorType
   | NumberToHexErrorType
   | FormatTransactionRequestErrorType
@@ -177,7 +161,7 @@ export async function call<TChain extends Chain | undefined>(
     const blockNumberHex = blockNumber ? numberToHex(blockNumber) : undefined
     const block = blockNumberHex || blockTag
 
-    const rpcStateOverride = parseStateOverride(stateOverride)
+    const rpcStateOverride = serializeStateOverride(stateOverride)
 
     const chainFormat = client.chain?.formatters?.transactionRequest?.format
     const format = chainFormat || formatTransactionRequest
@@ -367,71 +351,4 @@ export function getRevertErrorData(err: unknown) {
   if (!(err instanceof BaseError)) return undefined
   const error = err.walk() as RawContractError
   return typeof error?.data === 'object' ? error.data?.data : error.data
-}
-
-export type ParseStateMappingErrorType = InvalidBytesLengthErrorType
-
-export function parseStateMapping(
-  stateMapping: StateMapping | undefined,
-): RpcStateMapping | undefined {
-  if (!stateMapping || stateMapping.length === 0) return undefined
-  return stateMapping.reduce((acc, { slot, value }) => {
-    if (slot.length !== 66)
-      throw new InvalidBytesLengthError({
-        size: slot.length,
-        targetSize: 66,
-        type: 'hex',
-      })
-    if (value.length !== 66)
-      throw new InvalidBytesLengthError({
-        size: value.length,
-        targetSize: 66,
-        type: 'hex',
-      })
-    acc[slot] = value
-    return acc
-  }, {} as RpcStateMapping)
-}
-
-export type ParseAccountStateOverrideErrorType =
-  | NumberToHexErrorType
-  | StateAssignmentConflictErrorType
-  | ParseStateMappingErrorType
-
-export function parseAccountStateOverride(
-  args: Omit<StateOverride[number], 'address'>,
-): RpcAccountStateOverride {
-  const { balance, nonce, state, stateDiff, code } = args
-  const rpcAccountStateOverride: RpcAccountStateOverride = {}
-  if (code !== undefined) rpcAccountStateOverride.code = code
-  if (balance !== undefined)
-    rpcAccountStateOverride.balance = numberToHex(balance)
-  if (nonce !== undefined) rpcAccountStateOverride.nonce = numberToHex(nonce)
-  if (state !== undefined)
-    rpcAccountStateOverride.state = parseStateMapping(state)
-  if (stateDiff !== undefined) {
-    if (rpcAccountStateOverride.state) throw new StateAssignmentConflictError()
-    rpcAccountStateOverride.stateDiff = parseStateMapping(stateDiff)
-  }
-  return rpcAccountStateOverride
-}
-
-export type ParseStateOverrideErrorType =
-  | InvalidAddressErrorType
-  | AccountStateConflictErrorType
-  | ParseAccountStateOverrideErrorType
-
-export function parseStateOverride(
-  args?: StateOverride | undefined,
-): RpcStateOverride | undefined {
-  if (!args) return undefined
-  const rpcStateOverride: RpcStateOverride = {}
-  for (const { address, ...accountState } of args) {
-    if (!isAddress(address, { strict: false }))
-      throw new InvalidAddressError({ address })
-    if (rpcStateOverride[address])
-      throw new AccountStateConflictError({ address: address })
-    rpcStateOverride[address] = parseAccountStateOverride(accountState)
-  }
-  return rpcStateOverride
 }
